// packages/draw/core/src/types.ts
export enum ActionType {
    Connect = 'connect',
    AddNewBlock = 'newBlock',
    ChangePosition = 'changePosition',
    ChangeLabel = 'changeLabel',
    Highlight = 'highlight',
    RemoveHighlight = 'removeHighlight',
    HideBlock = 'hide',
    ChangeBlockStyle = 'changeBlockStyle',
    Show = 'show',
}
export interface BaseAction<TActionType extends ActionType, TActionValue extends ActionSetting> {
    name: TActionType
    value: TActionValue
}
export interface AddNewBlockAction extends BaseAction<ActionType.AddNewBlock, AddNewBlockActionSetting> {}

export interface ChangeBlockStyleAction extends BaseAction<ActionType.ChangeBlockStyle, BlockStyleActionSetting> {}
export interface ChangeLableAction extends BaseAction<ActionType.ChangeLabel, ChangeLabelActionSetting> {}
export interface ChangeBlockPositionAction extends BaseAction<ActionType.ChangePosition, ChangeBlockPositionActionSetting> {}
export interface ConnectAction extends BaseAction<ActionType.Connect, ConnectActionSetting> {}
export interface HideBlockAction extends BaseAction<ActionType.HideBlock, HideBlocksActionSetting> {}
export interface HighlightAction extends BaseAction<ActionType.Highlight, HighlightActionSetting> {}
export interface RemoveHighlightAction extends BaseAction<ActionType.RemoveHighlight, HideBlocksActionSetting> {}
export type Action =
    | AddNewBlockAction
    | ChangeLableAction
    | ChangeBlockPositionAction
    | ConnectAction
    | HideBlockAction
    | HighlightAction
    | ChangeBlockStyleAction
    | RemoveHighlightAction;
// ...
export enum ChangeType {
    AddNewBlock = 0,
    ChangeStyle = 1,
    RemoveBlock = 2,
    ChangePosition = 3,
}
export interface AddBlockChangeSetting {
    blockSettings: BlockGroupDescriptions | BlockElementDescription | null
    newBlockId: string
}
export interface ChangeBlockStyleSetting {
    blockId: string
    newStyle: Partial<BlockStyle>
}
export interface ChangeBlockLabelSetting {
    blockId: string
    label: string
}
export interface RemoveBlocksSetting {

}
// ...
export interface BaseChange {
    type: ChangeType
    owner?: ChangeOwner
}
export interface AddBlockChange extends BaseChange {
    type: ChangeType.AddNewBlock
    setting: AddBlockChangeSetting
}
export interface ChangeBlockStyle extends BaseChange {
    type: ChangeType.ChangeStyle
    setting: ChangeBlockStyleSetting
}
export interface RemoveBlocks extends BaseChange {
    type: ChangeType.RemoveBlock
    setting: RemoveBlocksSetting
}
export interface ChangePosition extends BaseChange {
    type: ChangeType.ChangePosition
    setting: {
        blockId: string
        x: number | FormulaValue
        y: number | FormulaValue
        w: number | FormulaValue
        h: number | FormulaValue
    }
}
export type Change = AddBlockChange | ChangeBlockStyle | RemoveBlocks | ChangePosition;